backend/requirements.txt
# Docling UI Backend Dependencies
flask==3.0.0
flask-cors==4.0.0
docling==2.14.0
sqlalchemy==2.0.36
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
werkzeug==3.0.1
gunicorn==21.2.0

